requirements.txt
errbot>=6.0.0
dnspython>=2.0.0
